+++ gui/packages/desktop/src/renderer/lib/notification-controller.js
@@ -7,9 +7,10 @@ import config from '../../config';
 import type { TunnelStateTransition } from './daemon-rpc';
 
 export default class NotificationController {
-  _activeNotification: ?Notification;
+  _lastTunnelStateNotification: ?Notification;
   _reconnecting = false;
   _presentedNotifications = {};
+  _pendingNotifications: Array<Notification> = [];
 
   notifyTunnelState(tunnelState: TunnelStateTransition) {
     switch (tunnelState.state) {
@@ -47,15 +48,24 @@ export default class NotificationController {
   }
 
   notifyInconsistentVersion() {
+    if (remote.getCurrentWindow().isVisible()) {
+      return;
+    }
+
     this._presentNotificationOnce('inconsistent-version', () => {
-      new Notification(remote.app.getName(), {
+      const notification = new Notification(remote.app.getName(), {
         body: 'Inconsistent internal version information, please restart the app',
         silent: true,
       });
+      this._addPendingNotification(notification);
     });
   }
 
   notifyUnsupportedVersion(upgradeVersion: string) {
+    if (remote.getCurrentWindow().isVisible()) {
+      return;
+    }
+
     this._presentNotificationOnce('unsupported-version', () => {
       const notification = new Notification(remote.app.getName(), {
         body: `You are running an unsupported app version. Please upgrade to ${upgradeVersion} now to ensure your security`,
@@ -65,33 +75,36 @@ export default class NotificationController {
       notification.addEventListener('click', () => {
         remote.shell.openExternal(config.links.download);
       });
+
+      this._addPendingNotification(notification);
     });
   }
 
+  cancelPendingNotifications() {
+    for (const notification of this._pendingNotifications) {
+      this._closeNotification(notification);
+    }
+  }
+
   _showTunnelStateNotification(message: string) {
-    const lastNotification = this._activeNotification;
+    const lastNotification = this._lastTunnelStateNotification;
     const sameAsLastNotification = lastNotification && lastNotification.body === message;
 
     if (sameAsLastNotification || remote.getCurrentWindow().isVisible()) {
       return;
     }
 
-    const newNotification = new Notification(remote.app.getName(), { body: message, silent: true });
-
-    this._activeNotification = newNotification;
-
-    newNotification.addEventListener('show', () => {
-      // If the notification is closed too soon, it might still get shown. If that happens, close()
-      // should be called again so that it is closed immediately.
-      // Tracking issue: https://github.com/electron/electron/issues/12887
-      if (this._activeNotification !== newNotification) {
-        newNotification.close();
-      }
+    const newNotification = new Notification(remote.app.getName(), {
+      body: message,
+      silent: true,
     });
 
     if (lastNotification) {
-      lastNotification.close();
+      this._closeNotification(lastNotification);
     }
+
+    this._lastTunnelStateNotification = newNotification;
+    this._addPendingNotification(newNotification);
   }
 
   _presentNotificationOnce(notificationName: string, presentNotification: () => void) {
@@ -99,6 +112,35 @@ export default class NotificationController {
     if (!presented[notificationName]) {
       presented[notificationName] = true;
       presentNotification();
+    }
+  }
+
+  _closeNotification(notification: Notification) {
+    // If the notification is closed too soon, it might still get shown. If that happens, close()
+    // should be called again so that it is closed immediately.
+    // Tracking issue: https://github.com/electron/electron/issues/12887
+    notification.addEventListener('show', () => {
+      notification.close();
+    });
+
+    notification.close();
+  }
+
+  _addPendingNotification(notification: Notification) {
+    // Quirk: chromium postpones the 'close' event until new notifications pump the queue or window
+    // becomes visible. It's possible that there is going to be one stale notification in
+    // `_pendingNotifications` array but that shouldn't be a big deal.
+    notification.addEventListener('close', () => {
+      this._removePendingNotification(notification);
+    });
+
+    this._pendingNotifications.push(notification);
+  }
+
+  _removePendingNotification(notification: Notification) {
+    const index = this._pendingNotifications.indexOf(notification);
+    if (index !== -1) {
+      this._pendingNotifications.splice(index, 1);
     }
   }
 }